pyproject.toml
[project]
name = "pcf-manifest"
version = "0.1.4"
description = "Author Power Apps Component Framework manifests as code and generate ControlManifest.Input.xml."
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Andrew Maspero", email = "andrew@vectorfy.co" }]
keywords = ["pcf", "power-apps", "power-platform", "manifest", "xml", "yaml"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Build Tools",
]
dependencies = [
  "pydantic>=2.6",
  "typer>=0.12",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=5.0",
  "playwright>=1.40",
]

[project.scripts]
pcf-manifest = "pcf_manifest_toolkit.cli:app"

[project.urls]
homepage = "https://github.com/andrewmaspero/pcf-manifest-toolkit"
repository = "https://github.com/andrewmaspero/pcf-manifest-toolkit"
issues = "https://github.com/andrewmaspero/pcf-manifest-toolkit/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pcf_manifest_toolkit.data" = ["*.json"]
